LiteHF.jl

Documentation for LiteHF.jl

LiteHF.FlatPrior — Type

Pseudo flat prior in the sense that `logpdf()` always evaluates to zero,
but `rand()`, `minimum()`, and `maximum()` behaves like `Uniform(a, b)`.

LiteHF.Histosys — Type

Histosys is defined by two vectors represending bin counts in hi_data and lo_data

LiteHF.Lumi — Type

Luminosity doesn't need interpolation, σ is provided at modifier construction time. In pyhf JSON, this information lives in the "Measurement" section, usually near the end of the JSON file.

LiteHF.Normfactor — Type

Normfactor is unconstrained, so interp is just identity.

LiteHF.Normsys — Method

Normsys is defined by two multiplicative scalars

LiteHF.Shapefactor — Type

Shapefactor is unconstrained, so interp is just identity. Unlike Normfactor, this is per-bin

LiteHF.Shapesys — Type

Shapesys doesn't need interpolation, similar to Staterror

LiteHF.Staterror — Type

Staterror doesn't need interpolation, but it's a per-bin modifier. Information regarding which bin is the target is recorded in bintwoidentity.

The δ is the absolute yield uncertainty in each bin, and the relative uncertainty: δ / nominal is taken to be the σ of the prior, i.e. α ~ Normal(1, δ/nominal)

LiteHF.build_channel — Method

build_channel(rawjdict[:channels][1][:samples][2]) =>
Dict{String, ExpCounts}

LiteHF.build_modifier! — Method

build_modifier(rawjdict[:channels][1][:samples][2][:modifiers][1]) =>
<:AbstractModifier

LiteHF.build_modifier — Method

build_modifier(...[:modifiers][1][:data], Type) =>
<:AbstractModifier

LiteHF.build_pyhf — Method

build_pyhf(load_pyhfjson(path)) 
    -> expected::Function, priors::NamedTupleDist, priornames::Tuple{Symbol}

the expected(αs) is a function that takes vector or tuple of length N, where N is also the length of priors and priornames. In other words, the three objects returned are aligned.

The output of forward_model() is always a vector of length N, represending the expected bin counts when all parameters (αs) taking a set of specific values.

LiteHF.build_sample — Function

build_sample(rawjdict[:channels][1][:samples][2]) =>
ExpCounts

LiteHF.load_pyhfjson — Method

load_pyhfjson(path)

LiteHF.pyhf_loglikelihoodof — Method

pyhf_loglikelihoodof(expected, obs, priors)

Return a callable Function that would calculate the log likelihood FlatPrior prior shouldn't have contribution to constraint

Note

The "constraint" terms that come from prior IS included here.

LiteHF.pyhf_loglikelihoodof — Method

pyhf_loglikelihoodof(expected, obs)

Return a callable Function that would calculate the log likelihood

Note

The "constraint" terms that come from prior is NOT included here.

LiteHF.twoidentity — Method

Identity function that takes two argument. Used for interpolation code that
doesn't do any interpolation.